New Addon RSVP-Close-Period
Allows admins to add a closing period for bookings in amount of hours.

For example:

Set a RSVP Close period of 12 hours.

Now, nobody will be able to book that event if the current time is within 12hrs of the event.

So event date/time = 27-07-2018/22:00

No bookings will be allowed after 10:00 on 27-07-2018